CRUD working for location conditions

+++ clock/views.py
@@ -1,12 +1,13 @@
 from .models import Clock, LocationCondition, CurrentState, State, UserProfile
-from .forms import LoginForm, ClockForm, StateForm
+from .forms import LoginForm, ClockForm, StateForm, LocationConditionForm
 from django.contrib.auth import authenticate, login, logout
 from geopy.distance import great_circle
 from django.utils import timezone
 from django.views.decorators.csrf import csrf_exempt
 from django.http import HttpResponse
 from django.core import serializers
-from django.shortcuts import render, redirect
+from django.shortcuts import render, redirect, get_object_or_404
+from django.views.generic import ListView
 from django.forms import formset_factory, modelformset_factory
 
 
@@ -72,12 +73,57 @@ def clock_form(request, clock_id=None):
         return render(request, 'clock/sign-in.html', context)
 
 
+def location_condition_form(request, clock_id, location_condition_id=None):
+    new_location_condition = location_condition_id is None
+    clock = Clock.objects.get(id=clock_id)
+    if request.method == 'POST':
+        if new_location_condition:
+            form = LocationConditionForm(request.POST)
+        else:
+            location_condition = LocationCondition.objects.get(
+                id=location_condition_id)
+            form = LocationConditionForm(
+                request.POST, instance=location_condition)
+        if form.is_valid():
+            location_condition = form.save(commit=False)
+            location_condition.radius_miles = request.POST.get('radius_miles')
+            location_condition.latitude = request.POST.get('latitude')
+            location_condition.longitude = request.POST.get('longitude')
+            if new_location_condition:
+                requester_profile = UserProfile.objects.get(
+                    user=request.user)
+                location_condition.user_profile = requester_profile
+            location_condition.save()
+        return redirect('/clock/{id}/location-conditions/'.format(id=clock_id))
+    else:
+        if new_location_condition:
+            form = LocationConditionForm()
+        else:
+            location_condition = LocationCondition.objects.get(
+                id=location_condition_id)
+            form = LocationConditionForm(instance=location_condition)
+        form.fields['state'].queryset = State.objects.filter(clock=clock)
+        return render(request,
+                      'clock/location-condition-form.html',
+                      {'form': form, 'clock': clock})
+
+
 def delete_clock(request):
     if request.user.is_authenticated and request.method == 'POST':
         clock = Clock.objects.get(id=request.POST.get('clock_id'))
         if clock.has_permission(request.user.username):
             clock.delete()
             return redirect('/')
+    return HttpResponse(status=550)
+
+
+def delete_location_condition(request):
+    if request.user.is_authenticated and request.method == 'POST':
+        location_condition = LocationCondition.objects.get(
+            id=request.POST.get('location_condition_id'))
+        if (location_condition.user_profile.user == request.user):
+            location_condition.delete()
+            return redirect('/clock/{id}/location-conditions/'.format(id=request.POST.get('clock_id')))
     return HttpResponse(status=550)
 
 
@@ -96,11 +142,17 @@ def manage_states(request, clock_id):
         formset = StateFormSet(request.POST, request.FILES)
         if formset.is_valid():
             for form in formset:
-                if form.is_valid() and not form.empty_permitted:
-                    form.instance.clock = clock
-                    form.save()
+                if form.is_valid():
+                    if form.empty_permitted:
+                        if form.instance.name and form.instance.position is not None:
+                            state = form.save(commit=False)
+                            state.clock = clock
+                            state.save()
+                    else:
+                        form.save()
             for form in formset.deleted_forms:
-                form.instance.delete()
+                if form.instance.id:
+                    form.instance.delete()
     return redirect('/clock/{id}'.format(id=clock_id))
 
 
@@ -140,3 +192,18 @@ def login_user(request):
 def logout_user(request):
     logout(request)
     return redirect('index')
+
+
+class LocationConditionList(ListView):
+    context_object_name = 'location_conditions'
+    template_name = 'clock/location-condition-list.html'
+
+    def get_queryset(self):
+        self.clock = get_object_or_404(Clock, id=self.kwargs['clock_id'])
+        return LocationCondition.objects.filter(state__clock=self.clock)
+
+    def get_context_data(self, **kwargs):
+        context = super(LocationConditionList, self).get_context_data(**kwargs)
+        # Add in the publisher
+        context['clock'] = self.clock
+        return context
